CCS-SYSTEM-main/backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
pymongo==4.8.0
mongomock==4.1.2
setuptools==80.10.2
Werkzeug==3.0.1
cryptography==41.0.7
bcrypt==4.1.2
flask-jwt-extended==4.6.0

